Rename file cbird to obj for more consistent module naming
