+++ catbird/__init__.py
@@ -1,5 +1,3 @@
-
-
-from .cbird import *
+from .obj import *
 from .model import *
 from .collection import *
